Changed default text of top ten GUI
As I thought that the default appearance of the text of the top ten heads was a little bland, I decided to change it a bit to make it a little more appealing. Just my suggestion, in no way do you have to accept this change.

If possible though, could you compile a version of the plugin with just this class edited? Would be much appreciated :)