Fixes tastybento/ASkyBlock-Bugs-N-Features#91
The issue was that permissions were being given to the player in the
world they were in before teleporting. For some reason the perm
sometimes could be interpreted as world-specific. Maybe this is
dependent on the permissions plugin. I changed the perms so that they
were checked, added and removed for the island world(s) specifically.