Fixes tastybento/ASkyBlock-Bugs-N-Features#104 AcidIsland inventory bug
The default island for AcidIsland was not showing titles, not resetting
money or inventory and not running commands on a new island. This is
because it is not pasted as a schematic and instead built up block by
block. The bug was introduced in 3.0.6.2 because the player reset code
was moved to the schematic class to make sure it occurred after
teleporting to the new island.