Update TopTen.java
I don't think this is going to be accepted so far, but it's worth trying it out. xD
You should also display the island owner, even if he doesn't have a team. If the island has a custom name, the name of the owner wouldn't be identifiable. So we should display the island owner too.

(This text was translated by Google Translate. Sometimes my english isn't that well :/ lol)