Fixes issue where the spawn block for the partner island was ignored
Addresses
https://github.com/tastybento/ASkyBlock-Bugs-N-Features/issues/60

Note, the spawn position is NOT checked for safety, so if you stick it
above lava or high up, the player will fall into it. Also, if the player
subsequently makes it an unsafe position, they will die. If a portal
exists, the player will teleport to the portal instead.