Fix to avoid team leaders losing their island.
If an ex-team mate's player file is wrong then their ex-leader would
lose their island if it had moved to a different location when the
ex-team member logs in.

This change will fix for both owners and ex-team mates. If an ex-team
mate has a bad file, then when logging in their file will be reset.


https://github.com/tastybento/ASkyBlock-Bugs-N-Features/issues/857#issuecomment-405056214